Set up timer driven complex events
Multiple images, random inter-image delays, it's all here. Unfortunately, the timer dispatching seems
to be causing a problem and the app eventually becomes unresponsive and if left after that runs out
of memory.

This is where I left off yesterday. Now I'll see if I can switch to using a setInterval based
architecture.